# compiler: make moduleContext.functions non-pointer slice