--- a/internal/engine/compiler/impl_amd64_test.go
+++ b/internal/engine/compiler/impl_amd64_test.go
@@ -22,7 +22,7 @@ func TestAmd64Compiler_indirectCallWithTargetOnCallingConvReg(t *testing.T) {
 	// and the typeID  matches the table[targetOffset]'s type ID.
 	operation := &wazeroir.OperationCallIndirect{TypeIndex: 0}
 	env.module().TypeIDs = []wasm.FunctionTypeID{0}
-	env.module().Engine = &moduleEngine{functions: []*function{}}
+	env.module().Engine = &moduleEngine{functions: []function{}}
 
 	me := env.moduleEngine()
 	{ // Compiling call target.
@@ -35,14 +35,14 @@ func TestAmd64Compiler_indirectCallWithTargetOnCallingConvReg(t *testing.T) {
 		c, _, err := compiler.compile()
 		require.NoError(t, err)
 
-		f := &function{
+		f := function{
 			parent:                &code{codeSegment: c},
 			codeInitialAddress:    uintptr(unsafe.Pointer(&c[0])),
 			moduleInstanceAddress: uintptr(unsafe.Pointer(env.moduleInstance)),
 			source:                &wasm.FunctionInstance{TypeID: 0},
 		}
 		me.functions = append(me.functions, f)
-		table[0] = uintptr(unsafe.Pointer(f))
+		table[0] = uintptr(unsafe.Pointer(&f))
 	}
 
 	compiler := env.requireNewCompiler(t, newCompiler, &wazeroir.CompilationResult{