Ref #889 

This further reduces the allocations during instantiation

```
$ benchstat old_re2.txt new_re2.txt
name                                    old time/op    new time/op    delta
Initialization/interpreter-10              136µs ± 2%     137µs ± 2%    ~     (p=0.222 n=5+5)
Initialization/interpreter-multiple-10    62.5µs ±17%    67.4µs ±13%    ~     (p=0.421 n=5+5)
Initialization/compiler-10                 167µs ± 5%     156µs ± 4%  -6.66%  (p=0.008 n=5+5)
Initialization/compiler-multiple-10       44.1µs ± 8%    43.0µs ± 9%    ~     (p=0.548 n=5+5)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-10              550kB ± 0%     550kB ± 0%    ~     (p=0.881 n=5+5)
Initialization/interpreter-multiple-10     551kB ± 0%     551kB ± 0%    ~     (p=0.548 n=5+5)
Initialization/compiler-10                 506kB ± 0%     498kB ± 0%  -1.62%  (p=0.008 n=5+5)
Initialization/compiler-multiple-10        507kB ± 0%     498kB ± 0%  -1.60%  (p=0.008 n=5+5)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-10              1.88k ± 0%     1.88k ± 0%    ~     (all equal)
Initialization/interpreter-multiple-10     1.88k ± 0%     1.88k ± 0%    ~     (all equal)
Initialization/compiler-10                  37.0 ± 0%      36.0 ± 0%  -2.70%  (p=0.008 n=5+5)
Initialization/compiler-multiple-10         44.0 ± 0%      43.0 ± 0%  -2.27%  (p=0.008 n=5+5)
```